Запушила сразу весь проект в репозиторий, поэтому pull request вот та…

+++ app/src/main/java/com/example/weathersecond/WeatherData/Main.java
@@ -1,31 +1,31 @@
 package com.example.weathersecond.WeatherData;
 
 public class Main {
-    private float temp;
-    private int pressure;
-    private int humidity;
+    private double temp;
+    private double pressure;
+    private double humidity;
 
-    public float getTemp() {
+    public double getTemp() {
         return temp;
     }
 
-    public void setTemp(float temp) {
+    public void setTemp(double temp) {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public double getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(double pressure) {
         this.pressure = pressure;
     }
 
-    public int getHumidity() {
+    public double getHumidity() {
         return humidity;
     }
 
-    public void setHumidity(int humidity) {
+    public void setHumidity(double humidity) {
         this.humidity = humidity;
     }
 }
